moving useComponentDidMount to common/hooks
